+++ sqlalchemy_drill/drilldbapi/exceptions.py
